Moved the loadout builders reducer to its own file to make things a little cleaner.
